+++ arangod/Graph/Providers/ClusterProvider.h
@@ -26,7 +26,7 @@
 
 #include "Graph/EdgeDocumentToken.h"
 #include "Graph/Providers/BaseProviderOptions.h"
-#include "Graph/Providers/BaseStep.h"
+#include "Graph/Steps/BaseStep.h"
 #include "Graph/Providers/TypeAliases.h"
 
 #include "Aql/TraversalStats.h"
