+++ arangod/Graph/Steps/BaseStep.h
@@ -25,7 +25,7 @@
 
 #include "Basics/ResultT.h"
 
-#include <velocypack/HashedStringRef.h>
+#include "velocypack/HashedStringRef.h"
 #include "Basics/ResultT.h"
 
 #include <numeric>
