+++ arangod/Graph/Steps/ClusterProviderStep.h
@@ -26,7 +26,7 @@
 
 #include <utility>
 
-#include "Graph/Providers/BaseStep.h"
+#include "BaseStep.h"
 #include "Graph/Providers/TypeAliases.h"
 #include "Transaction/Methods.h"
