+++ arangod/Graph/Steps/SingleServerProviderStep.h
@@ -24,7 +24,7 @@
 
 #include "Transaction/Methods.h"
 #include "Graph/EdgeDocumentToken.h"
-#include "Graph/Providers/BaseStep.h"
+#include "BaseStep.h"
 #include "Graph/Providers/TypeAliases.h"
 
 #pragma once
